Kepler Orbits
Add Simple Kepler Orbits package. Sun now now rotates on a simulated orbit.
NOTE: Moving the sun in the editor will change its orbintal relationship with the Earth. Be are of orbis when manually moving planets.

Package documentation (little of it) and example resources are still included in the project for reference and troubleshooting. They can be removed any time, and also reimported without issue when needed.